+++ src/formats/Molfile.cpp
@@ -169,7 +169,7 @@ template <MolfileFormat F> void Molfile<F>::read(Frame& frame) {
     }
 
     if (topology_) {
-        frame.resize(topology_->natoms());
+        frame.resize(topology_->size());
         frame.set_topology(*topology_);
     }
     molfile_to_frame(timestep, frame);