+++ CHANGELOG.md
@@ -19,6 +19,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Functions taking an atomic index parameter can now throw `OutOfBounds` errors
   if the index is out of bounds.
 * `Topology::append` is now called `Topology::add_atom`
+* `Topology::natoms` and `Frame::natoms` are now called `Topology::size` and
+  `Frame::size`
 * Added `Frame::distance`, `Frame::angle`, `Frame::dihedral` and
   `Frame::out_of_plane` to get geometric information on the system, accounting
   for periodic boundary conditions.